add - user routes done and tested

+++ routes/users.py
@@ -1,7 +1,10 @@
 import flask
 from flask import jsonify, abort, request, Blueprint
+from werkzeug.security import generate_password_hash
 
 import db.couchDB_service as db
+from models.User import User
+from routes.authentication import token_required
 
 users_route = Blueprint('users-route', __name__)
 
@@ -11,50 +14,62 @@ def get_blueprint():
     return users_route
 
 
-# Not quite done, must send to_public()
+# done
 @users_route.route('/', methods=['GET'])
-def get_all():
-    return jsonify(db.get_users())
+@token_required
+def get_all(current_user):
+    return jsonify(db.get_users()) if current_user['is_admin'] \
+        else flask.abort(401)
 
 
-# Done -> check token
+# Done
 @users_route.route('/<string:_id>', methods=['GET'])
-def get_with_id(_id):
+@token_required
+def get_with_id(current_user, _id):
     user = db.get_user_by_id(_id)
-    return jsonify(user.to_public()) if user else flask.abort(404)
+    if user:
+        return jsonify(user.to_admin()) if current_user['is_admin'] \
+            else jsonify(user.to_public())
+    else:
+        return flask.abort(404)
 
 
+# done
 @users_route.route('/<string:_id>/ban', methods=['POST'])
-def ban(_id):
-    if not request.get_json():
-        return jsonify(db.ban_user(None))
+@token_required
+def ban(current_user, _id):
+    if not current_user['is_admin']:
+        flask.abort(401)
 
-    data = request.get_json(force=True)
-
-    return jsonify(db.ban_user(data))
-
-
-# Done -> mais à checker
-@users_route.route('/', methods=['POST'])
-def create_one():
-    if not request.get_json():
-        abort(400)
-
-    data = request.get_json(force=True)
-    print(data)
-    return jsonify(db.create_user(data))
+    return jsonify(db.ban_user(_id))
 
 
+# done
 @users_route.route('/<string:_id>', methods=['DELETE'])
-def delete_one(_id):
-    return jsonify(db.delete_user(_id))
+@token_required
+def delete_one(current_user, _id):
+    return jsonify(db.delete_user(_id)) if current_user['is_admin'] \
+        else flask.abort(401)
 
 
+# done
 @users_route.route('/<string:_id>', methods=['PUT'])
-def edit_one(_id):
-    if not request.get_json():
-        abort(400)
+@token_required
+def edit_one(current_user, _id):
+    if current_user['_id'] != _id:
+        abort(401)
 
-    # data = request.get_json(force=True)
-
-    return jsonify(db.edit_user())
+    data = request.form
+    email = data.get('email')
+    first_name = data.get('first_name')
+    last_name = data.get('last_name')
+    campus = data.get('campus')
+    password = data.get('password')
+    user = User(_id=_id,
+                first_name=first_name,
+                last_name=last_name,
+                campus=campus,
+                email=email,
+                password=generate_password_hash(password)
+                )
+    return jsonify(db.edit_user(user, _id))
